Fix a bunch of warnings from PVS-Studio
