+++ src/Compression/tests/gtest_compressionCodec.cpp
@@ -35,29 +35,6 @@
 
 using namespace DB;
 
-namespace std
-{
-template <typename T>
-std::ostream & operator<<(std::ostream & ostr, const std::optional<T> & opt)
-{
-    if (!opt)
-    {
-        return ostr << "<empty optional>";
-    }
-
-    return ostr << *opt;
-}
-
-template <typename T>
-std::vector<T> operator+(std::vector<T> && left, std::vector<T> && right)
-{
-    std::vector<T> result(std::move(left));
-    std::move(std::begin(right), std::end(right), std::back_inserter(result));
-
-    return result;
-}
-
-}
 
 namespace
 {
@@ -337,6 +314,14 @@ CodecTestSequence operator+(CodecTestSequence && left, const CodecTestSequence &
     return left.append(right);
 }
 
+std::vector<CodecTestSequence> operator+(const std::vector<CodecTestSequence> & left, const std::vector<CodecTestSequence> & right)
+{
+    std::vector<CodecTestSequence> result(std::move(left));
+    std::move(std::begin(right), std::end(right), std::back_inserter(result));
+
+    return result;
+}
+
 template <typename T>
 CodecTestSequence operator*(CodecTestSequence && left, T times)
 {
@@ -362,7 +347,7 @@ std::ostream & operator<<(std::ostream & ostr, const Codec & codec)
 {
     return ostr << "Codec{"
                 << "name: " << codec.codec_statement
-                << ", expected_compression_ratio: " << codec.expected_compression_ratio
+                << ", expected_compression_ratio: " << *codec.expected_compression_ratio
                 << "}";
 }
 
@@ -775,15 +760,13 @@ auto FFand0Generator = []()
     return [step = 0](auto i) mutable
     {
         decltype(i) result;
-        if (step++ % 2 == 0)
-        {
-            memset(&result, 0, sizeof(result));
-        }
-        else
-        {
-            memset(&result, 0xFF, sizeof(result));
-        }
 
+        if (step % 2 == 0)
+            memset(&result, 0, sizeof(result));
+        else
+            memset(&result, 0xFF, sizeof(result));
+
+        ++step;
         return result;
     };
 };
