+++ src/DataTypes/tests/gtest_DataType_deserializeAsText.cpp
@@ -15,8 +15,6 @@
 
 #include <Core/iostream_debug_helpers.h>
 
-namespace std
-{
 
 template <typename T>
 inline std::ostream& operator<<(std::ostream & ostr, const std::vector<T> & v)
@@ -27,8 +25,6 @@ inline std::ostream& operator<<(std::ostream & ostr, const std::vector<T> & v)
         ostr << i << ", ";
     }
     return ostr << "] (" << v.size() << ") items";
-}
-
 }
 
 using namespace DB;
