+++ src/Functions/isValidUTF8.cpp
@@ -277,7 +277,7 @@ SOFTWARE.
             len -= 16;
         };
 
-        while (len >= 16) // NOLINT
+        while (len >= 16) // NOLINT //-V1044
             check_packed(_mm_loadu_si128(reinterpret_cast<const __m128i *>(data)));
 
         /// 0 <= len <= 15 for now. Reading data from data - 1 because of right padding of 15 and left padding
