+++ src/Interpreters/InterpreterWatchQuery.cpp
@@ -77,7 +77,7 @@ BlockIO InterpreterWatchQuery::execute()
     if (IBlockInputStream * stream = dynamic_cast<IBlockInputStream *>(streams[0].get()))
     {
         StreamLocalLimits limits;
-        limits.mode = LimitsMode::LIMITS_CURRENT;
+        limits.mode = LimitsMode::LIMITS_CURRENT; //-V1048
         limits.size_limits.max_rows = settings.max_result_rows;
         limits.size_limits.max_bytes = settings.max_result_bytes;
         limits.size_limits.overflow_mode = settings.result_overflow_mode;
