+++ src/Processors/Executors/PipelineExecutingBlockInputStream.h
@@ -26,7 +26,6 @@ public:
     void setProcessListElement(QueryStatus * elem) final;
     void setLimits(const StreamLocalLimits & limits_) final;
     void setQuota(const std::shared_ptr<const EnabledQuota> & quota_) final;
-    void addTotalRowsApprox(size_t value) final;
 
 protected:
     void readPrefixImpl() override;
