+++ src/Processors/QueryPlan/QueryPlan.cpp
@@ -158,7 +158,7 @@ QueryPipelinePtr QueryPlan::buildQueryPipeline(
         if (last_pipeline)
         {
             frame.pipelines.emplace_back(std::move(last_pipeline));
-            last_pipeline = nullptr;
+            last_pipeline = nullptr; //-V1048
         }
 
         size_t next_child = frame.pipelines.size();
