+++ src/Storages/MergeTree/MergeTreeDataPartWriterCompact.cpp
@@ -114,8 +114,8 @@ void writeColumnSingleGranule(
     ISerialization::SerializeBinaryBulkSettings serialize_settings;
 
     serialize_settings.getter = stream_getter;
-    serialize_settings.position_independent_encoding = true;
-    serialize_settings.low_cardinality_max_dictionary_size = 0;
+    serialize_settings.position_independent_encoding = true; //-V1048
+    serialize_settings.low_cardinality_max_dictionary_size = 0; //-V1048
 
     serialization->serializeBinaryBulkStatePrefix(serialize_settings, state);
     serialization->serializeBinaryBulkWithMultipleStreams(*column.column, from_row, number_of_rows, serialize_settings, state);
